demo-storage/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.2
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
